src/requirements.txt
click==8.1.3
colorama==0.4.5
firebird-base==1.4.1
firebirdsql==1.2.2
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
protobuf==4.21.6
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
toml==0.10.2
Werkzeug==2.2.2
zipp==3.8.1
